requirements.txt
Pillow~=10.3.0
newspaper3k
urllib3
streamlit~=1.33.0
beautifulsoup4~=4.13.0b2
protobuf==3.20.0
jinja2>=3.0

bs4~=0.0.2
nltk~=3.8.1
requests~=2.31.0